Refactor and use 8 bit integers only

+++ tictactoe/src/lib/brick.rs
@@ -2,7 +2,7 @@ use crate::lib::player::Player;
 
 #[derive(Copy, Clone, Debug)]
 pub struct Brick {
-    pub x: i32,
-    pub y: i32,
+    pub x: i8,
+    pub y: i8,
     pub player: Player,
 }
